created new VideoContent componet, updates to FAQ
